+++ libs/providers/launchdarkly-client/src/lib/launchdarkly-client-provider.ts
@@ -3,14 +3,21 @@ import {
   Provider,
   JsonValue,
   ResolutionDetails,
-  StandardResolutionReasons, ErrorCode, ProviderMetadata, Logger
+  StandardResolutionReasons,
+  ErrorCode,
+  ProviderMetadata,
+  Logger,
+  GeneralError,
+  OpenFeatureEventEmitter,
+  ProviderEvents,
+  ProviderStatus
 } from '@openfeature/web-sdk';
-import {
-  basicLogger, LDClient
-} from 'launchdarkly-js-client-sdk';
-import isEqual from 'lodash.isequal';
 
-import {LaunchDarklyProviderOptions} from './launchdarkly-provider-options';
+import isEmpty from 'lodash.isempty';
+
+import { basicLogger, LDClient, initialize, LDOptions, LDFlagChangeset } from 'launchdarkly-js-client-sdk';
+
+import { LaunchDarklyProviderOptions } from './launchdarkly-provider-options';
 import translateContext from './translate-context';
 import translateResult from './translate-result';
 
@@ -34,28 +41,78 @@ export class LaunchDarklyClientProvider implements Provider {
     name: 'launchdarkly-client-provider',
   };
 
+  private readonly ldOptions: LDOptions | undefined;
   private readonly logger: Logger;
+  private _client?: LDClient;
 
-  constructor(private client: LDClient, options: LaunchDarklyProviderOptions = {}) {
-    if (options.logger) {
-      this.logger = options.logger;
+  public events = new OpenFeatureEventEmitter();
+
+  /*
+   * implement status field/accessor
+   * https://openfeature.dev/specification/sections/providers#requirement-242
+   * */
+  private _status: ProviderStatus = ProviderStatus.NOT_READY;
+
+  set status(status: ProviderStatus) {
+    this._status = status;
+  }
+
+  get status() {
+    return this._status;
+  }
+
+  constructor(
+    private readonly envKey: string,
+    { logger, ...ldOptions }: LaunchDarklyProviderOptions) {
+    if (logger) {
+      this.logger = logger;
     } else {
       this.logger = basicLogger({ level: 'info' });
     }
+    this.ldOptions = { ...ldOptions, logger: this.logger };
+
   }
 
-  async initialize(context: EvaluationContext): Promise<void> {
-    const newContext = this.translateContext(context);
-    const oldContext = this.client.getContext();
-    const ignoreAnonymous = newContext.anonymous && oldContext.anonymous && !newContext.key
-    if(!ignoreAnonymous && !isEqual(newContext, oldContext)) {
-       await this.client.identify(newContext);
+  private get client(): LDClient {
+    if (!this._client) {
+      throw new GeneralError('Provider is not initialized');
     }
-    return this.client.waitUntilReady();
+    return this._client;
+  }
+
+  async initialize(context?: EvaluationContext): Promise<void> {
+    const _context = isEmpty(context) ? { anonymous: true } : this.translateContext(context);
+    this._client = initialize(this.envKey, _context, this.ldOptions);
+
+    /*
+     * set event listeners to propagate ld events to open feature provider,
+     * use LD property streaming, since when enabled you are opting in for straming updates
+     * */
+    if (this.ldOptions?.streaming) {
+      this.setListeners();
+    }
+
+    try {
+      await this._client.waitForInitialization();
+      this.status = ProviderStatus.READY;
+    } catch {
+      this.status = ProviderStatus.ERROR;
+    }
   }
 
   onClose(): Promise<void> {
     return this.client.close();
+  }
+
+  /** set listeners to LD client and event the correspodent event in the Provider
+   * necessary for LD streaming changes
+   * */
+  private setListeners() {
+    this.client.on('change', (changeset: LDFlagChangeset) => {
+      this.events.emit(ProviderEvents.ConfigurationChanged, {
+        flagsChanged: Object.keys(changeset),
+      });
+    });
   }
 
   async onContextChange(oldContext: EvaluationContext, newContext: EvaluationContext): Promise<void> {
@@ -64,10 +121,7 @@ export class LaunchDarklyClientProvider implements Provider {
   }
 
   resolveBooleanEvaluation(flagKey: string, defaultValue: boolean): ResolutionDetails<boolean> {
-    const res = this.client.variationDetail(
-      flagKey,
-      defaultValue,
-    );
+    const res = this.client.variationDetail(flagKey, defaultValue);
     if (typeof res.value === 'boolean') {
       return translateResult(res);
     }
@@ -75,10 +129,7 @@ export class LaunchDarklyClientProvider implements Provider {
   }
 
   resolveNumberEvaluation(flagKey: string, defaultValue: number): ResolutionDetails<number> {
-    const res = this.client.variationDetail(
-      flagKey,
-      defaultValue,
-    );
+    const res = this.client.variationDetail(flagKey, defaultValue);
     if (typeof res.value === 'number') {
       return translateResult(res);
     }
@@ -86,10 +137,7 @@ export class LaunchDarklyClientProvider implements Provider {
   }
 
   resolveObjectEvaluation<T extends JsonValue>(flagKey: string, defaultValue: T): ResolutionDetails<T> {
-    const res = this.client.variationDetail(
-      flagKey,
-      defaultValue,
-    );
+    const res = this.client.variationDetail(flagKey, defaultValue);
     if (typeof res.value === 'object') {
       return translateResult(res);
     }
@@ -97,10 +145,7 @@ export class LaunchDarklyClientProvider implements Provider {
   }
 
   resolveStringEvaluation(flagKey: string, defaultValue: string): ResolutionDetails<string> {
-    const res = this.client.variationDetail(
-      flagKey,
-      defaultValue,
-    );
+    const res = this.client.variationDetail(flagKey, defaultValue);
     if (typeof res.value === 'string') {
       return translateResult(res);
     }
@@ -110,5 +155,4 @@ export class LaunchDarklyClientProvider implements Provider {
   private translateContext(context: EvaluationContext) {
     return translateContext(this.logger, context);
   }
-
 }